+++ AmazonChimeSDK/AmazonChimeSDK/device/MediaDevice.swift
@@ -33,6 +33,12 @@ import Foundation
     static func fromVideoDevice(device: VideoDevice?) -> MediaDevice {
         return MediaDevice(label: device?.name ?? "unknown", videoDevice: device)
     }
+    
+    public init(label: String, type: MediaDeviceType) {
+        self.label = label
+        self.type = type
+        self.port = nil
+    }
 
     public init(label: String, port: AVAudioSessionPortDescription? = nil, videoDevice: VideoDevice? = nil) {
         self.label = label
@@ -54,6 +60,8 @@ import Foundation
                 type = .audioHandset
             case .headphones, .headsetMic:
                 type = .audioWiredHeadset
+            case .builtInSpeaker:
+                type = .audioBuiltInSpeaker
             default:
                 type = .other
             }